✨🆕 updated

+++ main.py
@@ -5,28 +5,41 @@ import random
 import string
 
 def test(file="main.cpp"):
-    judge.compile(file)
-    tests=os.listdir("tests")
+    if judge.compile(file):
+        judge.judge(1)
+        tests=os.listdir("tests")
+        total_time=0
+        total_score=0
+        score=100/len(tests)
+        details=[]
+        results=[]
 
-    total_time=0
-    total_score=0
-    score=100/len(tests)
-
-    for i in range(len(tests)):
-        values=judge.judge(i+1).split("|")
-        result=values[0]
-        time=values[1]
-        st_score=0.0
-        if result=="PASS":
-            total_score+=score
-            st_score=score
-        print("#"+str(i+1)+": "+str(result)+" ["+str("{:.2f}".format(st_score))+"/"+str("{:.2f}".format(score))+"] ("+str(round(float(time)*1000))+"ms)")
-        total_time+=round(float(time)*1000)
-
-    total_score="{:.2f}".format(total_score)
-    print("Score: "+str(total_score)+"/100.00")
-    print("Time: "+str(int(total_time))+"ms")
-    os.remove("files/c++")
+        for i in range(len(tests)):
+            values=judge.judge(i+1).split("|")
+            result=values[0]
+            time=values[1]
+            st_score=0.0
+            if result=="PASS":
+                total_score+=score
+                st_score=score
+                results.append("P")
+            elif result=="FAILED":
+                results.append("-")
+            elif result=="Time limit exceeded":
+                results.append("T")
+            else:
+                results.append("X")
+            details.append("#"+str(i+1)+": "+str(result)+" ["+str("{:.2f}".format(st_score))+"/"+str("{:.2f}".format(score))+"] ("+str(round(float(time)*1000))+"ms)")
+            total_time+=round(float(time)*1000)
+        os.remove("files/c++")
+        print("\n".join(details))
+        print("\n===RESULT===")
+        print("".join(results))
+        total_score="{:.2f}".format(total_score)
+        print("Score: "+str(total_score)+"/100.00")
+        print("Time: "+str(int(total_time))+"ms\n===RESULT===\n")
+    else:
+        print("Compilation error.")
 
 def addTestCase():
     tests=os.listdir("tests")
@@ -62,7 +75,11 @@ print("2. Add a test case")
 print("3. Delete all test cases (THIS CANNOT BE UNDONE)")
 option=input("> ")
 if option=="1":
-    test()
+    fileName=input("File name (Default: main.cpp): ")
+    if not fileName:
+        test()
+    else:
+        test(fileName)
 elif option=="2":
     addTestCase()
 elif option=="3":
